requirements.txt
Flask==1.1.2
flasgger==0.9.5
PyYAML==5.4.1
marshmallow==3.11.1
itsdangerous==1.1.0
attrs==20.3.0
pyserial==3.5
SQLAlchemy==1.4.12
bcrypt==3.2.0
pycryptodome==3.10.1
python-dotenv==0.17.0
pyduk @ git+https://github.com/ConsonantSpork/pyduk@0.1.1
alembic==1.6.2
Flask-Login==0.5.0
WTForms==2.3.3
click==7.1.2
pyduk==0.1.1